requirements.txt
psycopg2-binary >= 2.9 
qdrant-client >= 1.6
requests >= 2.31
flask >= 3.1.0